Byron.Spec.Chain.STS.Rule.Chain

Documentation

data CHAIN

isHeaderSizeTooBigFailure ∷ PredicateFailure CHAIN → Bool

headerIsValid ∷ UPIState → BlockHeader → Rule CHAIN Transition ()

disL ∷ Lens' (State CHAIN) DIState

Lens for the delegation interface state contained in the chain state.

data ShouldGenDelegation

Constructors

GenDelegation

NoGenDelegation

data ShouldGenUTxO

Constructors

GenUTxO

NoGenUTxO

data ShouldGenUpdate

Constructors

GenUpdate

NoGenUpdate

sigGenChain ∷ ShouldGenDelegation → ShouldGenUTxO → ShouldGenUpdate → Environment CHAIN → State CHAIN → Gen (Signal CHAIN)

tamperWithPayloadHash ∷ Block → Gen Block

Produce an invalid hash for one of the three types of block payloads:

Delegation

Update

UTxO

invalidProofsBlockGen ∷ SignalGenerator CHAIN

Generate a block in which one of the three types of payload hashes:

Delegation

Update

UTxO

is invalid.

coverInvalidBlockProofs

Arguments

∷ (MonadTest m, HasCallStack, Data a)

⇒ CoverPercentage
Minimum percentage that each failure must occur.

→ a
Structure containing the failures

→ m ()